Report job's arguments when using activejob with delayed job
Activejob serializes the jobs arguments in the payload under
arguments rather than args.
https://github.com/rails/rails/blob/3bfd35245278f410bb2165e859f1cfeb352e0d77/activejob/lib/active_job/core.rb#L85

Due to this failed jobs do not report their arguments to Bugsnag
when using Activejob with DelayedJob.

This includes the payloads arguments if they are set in the same
failsafe way as the other payload attributes are reported with the
Delayedjob integration